app_3/requirements.txt
paho-mqtt
python-dotenv
prettytable